Create common base class for components
