+++ src/fastmcp/prompts/prompt.py
@@ -5,13 +5,13 @@ from __future__ import annotations as _annotations
 import inspect
 from abc import ABC, abstractmethod
 from collections.abc import Awaitable, Callable, Sequence
-from typing import TYPE_CHECKING, Annotated, Any
+from typing import TYPE_CHECKING, Any
 
 import pydantic_core
 from mcp.types import EmbeddedResource, ImageContent, PromptMessage, Role, TextContent
 from mcp.types import Prompt as MCPPrompt
 from mcp.types import PromptArgument as MCPPromptArgument
-from pydantic import BeforeValidator, Field, TypeAdapter, validate_call
+from pydantic import Field, TypeAdapter, validate_call
 
 from fastmcp.exceptions import PromptError
 from fastmcp.server.dependencies import get_context
@@ -19,7 +19,7 @@ from fastmcp.utilities.json_schema import compress_schema
 from fastmcp.utilities.logging import get_logger
 from fastmcp.utilities.types import (
     FastMCPBaseModel,
-    _convert_set_defaults,
+    FastMCPComponent,
     find_kwarg_by_type,
     get_cached_typeadapter,
 )
@@ -66,25 +66,12 @@ class PromptArgument(FastMCPBaseModel):
     )
 
 
-class Prompt(FastMCPBaseModel, ABC):
+class Prompt(FastMCPComponent, ABC):
     """A prompt template that can be rendered with parameters."""
 
-    name: str = Field(description="Name of the prompt")
-    description: str | None = Field(
-        default=None, description="Description of what the prompt does"
-    )
-    tags: Annotated[set[str], BeforeValidator(_convert_set_defaults)] = Field(
-        default_factory=set, description="Tags for the prompt"
-    )
     arguments: list[PromptArgument] | None = Field(
         default=None, description="Arguments that can be passed to the prompt"
     )
-
-    def __eq__(self, other: object) -> bool:
-        if type(self) is not type(other):
-            return False
-        assert isinstance(other, type(self))
-        return self.model_dump() == other.model_dump()
 
     def to_mcp_prompt(self, **overrides: Any) -> MCPPrompt:
         """Convert the prompt to an MCP prompt."""
