+++ src/fastmcp/utilities/types.py
@@ -2,22 +2,53 @@
 
 import base64
 import inspect
-from collections.abc import Callable
+from collections.abc import Callable, Sequence
 from functools import lru_cache
 from pathlib import Path
 from types import UnionType
 from typing import Annotated, TypeVar, Union, get_args, get_origin
 
 from mcp.types import ImageContent
-from pydantic import BaseModel, ConfigDict, TypeAdapter
+from pydantic import BaseModel, BeforeValidator, ConfigDict, Field, TypeAdapter
 
 T = TypeVar("T")
+
+
+def _convert_set_default_none(maybe_set: set[T] | Sequence[T] | None) -> set[T]:
+    """Convert a sequence to a set, defaulting to an empty set if None."""
+    if maybe_set is None:
+        return set()
+    if isinstance(maybe_set, set):
+        return maybe_set
+    return set(maybe_set)
 
 
 class FastMCPBaseModel(BaseModel):
     """Base model for FastMCP models."""
 
     model_config = ConfigDict(extra="forbid")
+
+
+class FastMCPComponent(FastMCPBaseModel):
+    """Base class for FastMCP tools, prompts, resources, and resource templates."""
+
+    name: str = Field(
+        description="The name of the component.",
+    )
+    description: str | None = Field(
+        default=None,
+        description="The description of the component.",
+    )
+    tags: Annotated[set[str], BeforeValidator(_convert_set_default_none)] = Field(
+        default_factory=set,
+        description="Tags for the component.",
+    )
+
+    def __eq__(self, other: object) -> bool:
+        if type(self) is not type(other):
+            return False
+        assert isinstance(other, type(self))
+        return self.model_dump() == other.model_dump()
 
 
 @lru_cache(maxsize=5000)
@@ -78,15 +109,6 @@ def find_kwarg_by_type(fn: Callable, kwarg_type: type) -> str | None:
         if is_class_member_of_type(param.annotation, kwarg_type):
             return name
     return None
-
-
-def _convert_set_defaults(maybe_set: set[T] | list[T] | None) -> set[T]:
-    """Convert a set or list to a set, defaulting to an empty set if None."""
-    if maybe_set is None:
-        return set()
-    if isinstance(maybe_set, set):
-        return maybe_set
-    return set(maybe_set)
 
 
 class Image:
